% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{groupfsInf}
\alias{groupfsInf}
\title{Compute selective p-values for a model fitted by \code{groupfs}.}
\usage{
groupfsInf(obj, sigma = NULL, type = c("all", "aic"), ntimes = 2,
  verbose = FALSE)
}
\arguments{
\item{obj}{Object returned by \code{\link{groupfs}} function}

\item{sigma}{Estimate of error standard deviation. If NULL (default), this is estimated using the mean squared residual of the full least squares fit when n >= 2p, and the mean squared residual of the selected model when n < 2p. In the latter case, the user should use \code{\link{estimateSigma}} function for a more accurate estimate.}

\item{type}{Type of conditional p-values to compute. With "all" (default), p-values are computed conditional on the final model with all variables up to \code{maxsteps}; with "aic" the number of steps is chosen after which the AIC criterion increases \code{ntimes} in a row, and then the same type of analysis as in "all" is carried out for the active variables at that number of steps.}

\item{ntimes}{Number of steps for which AIC criterion has to increase before minimizing point is declared.}

\item{verbose}{Print out progress along the way? Default is FALSE.}
}
\value{
An object of class "groupfsInf" containing selective p-values for the fitted model \code{obj}. For comparison with \code{\link{fsInf}}, note that the option \code{type = "active"} is not available.

\describe{
  \item{vars}{Labels of the active groups in the order they were included.}
  \item{pv}{Selective p-values computed from appropriate truncated distributions.}
  \item{sigma}{Estimate of error variance used in computing p-values.}
  \item{TC}{Observed value of truncated chi.}
  \item{df}{Rank of group of variables when it was added to the model.}
  \item{support}{List of intervals defining the truncation region of the truncated chi.}
}
}
\description{
Computes p-values for each group of variables in a model fitted by \code{\link{groupfs}}. These p-values adjust for selection by truncating the usual \eqn{\chi^2} statistics to the regions implied by the model selection event. Details are provided in a forthcoming work.
}

