% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.groupfs.R
\name{predict.groupfs}
\alias{predict.groupfs}
\title{Prediction and coefficient functions for \code{\link{groupfs}}.}
\description{
Make predictions or extract coefficients from a groupfs forward stepwise object.}
\usage{
\method{predict}{groupfs}(object, newx)
}
\arguments{
\item{object}{Object returned by a call to \code{\link{groupfs}}.}

\item{newx}{Matrix of x values at which the predictions are desired. If NULL, the x values from groupfs fitting are used.}
}
\value{
A vector of predictions or a vector of coefficients.
}

