% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs.groupfs.R
\name{scaleGroups}
\alias{scaleGroups}
\title{Center and scale design matrix by groups}
\usage{
scaleGroups(x, index, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{Design matrix.}

\item{index}{Group membership indicator of length p.}

\item{center}{Center groups, default is TRUE.}

\item{scale}{Scale groups by Frobenius norm, default is TRUE.}
}
\value{
Scaled design matrix
}
\description{
Center and scale design matrix by groups
}

