% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs.groupfs.R
\name{groupfsInf}
\alias{groupfsInf}
\title{Compute selective p-values for a model fitted by \code{groupfs}.}
\usage{
groupfsInf(obj, sigma = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{Object returned by \code{\link{groupfs}} function}

\item{sigma}{Estimate of error standard deviation. If NULL (default), p-values will be computed from a selective F test.}

\item{verbose}{Print out progress along the way? Default is TRUE.}
}
\value{
An object of class "groupfsInf" containing selective p-values for the fitted model \code{obj}. For comparison with \code{\link{fsInf}}, note that the option \code{type = "active"} is not available.

\describe{
  \item{vars}{Labels of the active groups in the order they were included.}
  \item{pv}{Selective p-values computed from appropriate truncated distributions.}
  \item{sigma}{Estimate of error variance used in computing p-values.}
  \item{TC}{Observed value of truncated chi.}
  \item{df}{Rank of group of variables when it was added to the model.}
  \item{support}{List of intervals defining the truncation region of the truncated chi.}
}
}
\description{
Computes p-values for each group of variables in a model fitted by \code{\link{groupfs}}. These p-values adjust for selection by truncating the usual \eqn{\chi^2} statistics to the regions implied by the model selection event. Details are provided in a forthcoming work.
}

