% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs.groupfs.R
\name{groupfsInf}
\alias{groupfsInf}
\title{Form univariate selection interval for a given contrast}
\usage{
groupfsInf(obj, projs = NULL)
}
\arguments{
\item{obj}{Object returned by \code{\link{groupfs}} function}

\item{projs}{Additional projections to define model selection event. For use with cross-validation. Default is NULL and it is not recommended to change this.}

\item{sigma}{Estimate of error standard deviation. If NULL (default), this is estimated using the mean squared residual of the full least squares fit when n >= 2p, and the mean squared residual of the selected model when n < 2p. In the latter case, the user should use \code{\link{estimateSigma}} function for a more accurate estimate.}
}
\value{
An object of class "groupfsInf" containing selective p-values for the fitted model \code{obj}. The default printing behavior should supply adequate information.

\describe{
  \item{vars}{Labels of the active groups in the order they were included.}
  \item{pv}{Selective p-values computed from appropriate truncated distributions.}
  \item{sigma}{Estimate of error variance used in computing p-values.}
  \item{TC}{Observed value of truncated chi.}
  \item{df}{Rank of group of variables when it was added to the model.}
  \item{support}{List of intervals defining the truncation region of the truncated chi.}
}
}
\description{
Form univariate selection interval for a given contrast
}

