% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.groupfs.R
\name{scaleGroups}
\alias{scaleGroups}
\title{Center and scale design matrix by groups}
\usage{
scaleGroups(x, index, center = TRUE, normalize = TRUE)
}
\arguments{
\item{x}{Design matrix.}

\item{index}{Group membership indicator of length p.}

\item{center}{Center groups, default is TRUE.}

\item{normalize}{Scale groups by Frobenius norm, default is TRUE.}
}
\value{
\describe{
  \item{x}{Optionally centered/scaled design matrix.}
  \item{xm}{Means of groups in original design matrix.}
  \item{xs}{Frobenius norms of groups in original design matrix.}
}
}
\description{
For internal use by \code{\link{groupfs}}.
}

