\name{selectiveInference-internal}
\title{Internal PMA functions}
\alias{print.fixedLassoInf}
  \alias{print.fs} 
\alias{print.fsInf} 
\alias{print.larInf}
\alias{print.lar}
  \alias{print.manyMeans}



\description{Internal selectiveInference functions}
\usage{
\method{print}{fs}(x, ...)
\method{print}{fsInf}(x, tailarea = TRUE, ...)
\method{print}{lar}(x,...)
\method{print}{larInf}(x, tailarea = TRUE, ...)
\method{print}{fixedLassoInf}(x, tailarea = TRUE, ...)
\method{print}{manyMeans}(x,...)

}
\author{Ryan Tibshirani, Rob Tibshirani, Jonathan Taylor, Joshua Loftus, Stephen Reid}
\keyword{internal}

